Map and items
I set up the map. I also set up one of my NPCs. this NPC holds a key that the player needs. The main thing that i worked on is adding all the items to the game. I added all of the components for all three dishes to their appropriate places in the game.