Working on talking to NPC and combining items
My main focus of this class was to add more conversation to the NPC and work on the combination aspect of my game. I could not get the dialouge code that a wrote to work so I emailed Mrs. Kiang for help. I also asked her for help with my combination puzzle since it has more components than any other combination problem that i have seen in the past.