Fixing NPC and combination puzzle
I polished up the interactions with the NPC. I also added a few more components for the dishes. I got help from Mrs. Kiang and implemented a table system to make is easy to make the final dishes. the table seems to be working as intended besides one bug. Even if the player has all of the components for a dish inform wont make it.