adding some rules and fixing bugs
I fixed a few problems that i was having with some of the descriptions in the game. I also added some code to prevent the player from eating the completed dishes that they make. I started to work on the way that the player will complete the puzzles. I am going to have the payer place the finished dish on a table and ring a bell for the food to be picked up.